amazon/pull-orders/requirements.txt
python-amazon-sp-api
boto3
requests